Create README.md
At Rosedale Pet Resort, we believe that every pet is special and should be accepted and loved like they are when they are home! We also believe in quality over quantity. Our goal is to create a place where each of our guests receive the attention they need and deserve, and are kept safe while in our care.
 6015 Rosedale St NW, Gig Harbor, WA 98335